+++ messages.go
@@ -4,7 +4,7 @@ import (
 	"database/sql/driver"
 
 	"github.com/jackc/pgio"
-	"github.com/jackc/pgtype"
+	"github.com/khaibin/pgtype"
 )
 
 const (