+++ rows.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgproto3/v2"
-	"github.com/jackc/pgtype"
+	"github.com/khaibin/pgtype"
 )
 
 // Rows is the result set returned from *Conn.Query. Rows must be closed before