+++ stdlib/sql.go
@@ -64,7 +64,7 @@ import (
 	errors "golang.org/x/xerrors"
 
 	"github.com/jackc/pgconn"
-	"github.com/jackc/pgtype"
+	"github.com/khaibin/pgtype"
 	"github.com/jackc/pgx/v4"
 )